Implementação da captura de imagem da crise
