--- Alertas/alertaApp/src/main/java/br/ita/bditac/mobile/alertas/CaptureImageActivity.java
@@ -1,38 +1,0 @@
-package br.ita.bditac.mobile.alertas;
-
-import android.content.Intent;
-import android.graphics.Bitmap;
-import android.os.Bundle;
-import android.provider.MediaStore;
-import android.support.v7.app.AppCompatActivity;
-import android.view.View;
-import android.widget.ImageView;
-
-public class CaptureImageActivity extends AppCompatActivity {
-
-    private static final int CAMERA_REQUEST = 1000;
-
-    ImageView imageView;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_capture_image);
-
-        imageView = (ImageView) findViewById(R.id.camera_image);
-    }
-
-    public void Capture_Click(View view) {
-        Intent intent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-        if(intent.resolveActivity(getPackageManager()) != null) {
-            startActivityForResult(intent, CAMERA_REQUEST);
-        }
-    }
-
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        if(requestCode == CAMERA_REQUEST && resultCode == RESULT_OK) {
-            imageView.setImageBitmap((Bitmap) data.getExtras().get("data"));
-        }
-    }
-
-}
