+++ Alertas/alertaApp/src/main/res/layout/activity_cadastrar_crise.xml
@@ -17,7 +17,7 @@
         <TextView
             android:layout_width ="wrap_content"
             android:layout_height ="wrap_content"
-            android:text ="@string/crises_report" />
+            android:text ="@string/crises_report"/>
 
         <Spinner
             android:id ="@+id/categoria"
@@ -35,30 +35,49 @@
             android:lines="10"
             android:minLines="5"
             android:maxLines="20"
-            android:scrollbars="vertical"
-            />
+            android:scrollbars="vertical"/>
+
+        <TextView
+            android:layout_width="349dp"
+            android:layout_height="wrap_content"
+            android:text="@string/click_the_below_button_to_take_photo"
+            android:textAlignment="center"
+            android:paddingBottom="@dimen/activity_vertical_margin"
+            android:paddingLeft="@dimen/activity_horizontal_margin"
+            android:paddingRight="@dimen/activity_horizontal_margin"
+            android:paddingTop="@dimen/activity_vertical_margin"/>
+
+        <ImageView
+            android:id="@+id/camera_image"
+            android:layout_width="match_parent"
+            android:layout_height="300dp"
+            android:paddingBottom="@dimen/activity_vertical_margin"
+            android:paddingLeft="@dimen/activity_horizontal_margin"
+            android:paddingRight="@dimen/activity_horizontal_margin"
+            android:paddingTop="@dimen/activity_vertical_margin"/>
+
+        <Button
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="@string/take_photo"
+            android:layout_gravity="center_horizontal"
+            android:paddingBottom="@dimen/activity_vertical_margin"
+            android:paddingLeft="@dimen/activity_horizontal_margin"
+            android:paddingRight="@dimen/activity_horizontal_margin"
+            android:paddingTop="@dimen/activity_vertical_margin"
+            android:onClick="Capture_Click"/>
 
         <Button
             android:layout_width ="match_parent"
             android:layout_height ="wrap_content"
             android:text ="@string/save"
-            android:onClick="Salvar_Click"
-            />
+            android:onClick="Salvar_Click"/>
 
         <Button
             android:layout_width ="match_parent"
             android:layout_height ="wrap_content"
             android:text ="@string/clear"
-            android:onClick="Limpar_Click"
-            />
-
-        <Button
-            android:id ="@+id/image"
-            android:layout_width ="match_parent"
-            android:layout_height ="wrap_content"
-            android:text ="@string/image"
-            android:onClick="Imagem_Click"
-            />
+            android:onClick="Limpar_Click"/>
 
     </LinearLayout>
